rename a file and refactor correspondingly
